Organise populate files
